+++ src/main/java/com/deviget/minesweeper/infra/service/CellOperation.java
@@ -1,8 +1,25 @@
 package com.deviget.minesweeper.infra.service;
 
-public enum CellOperation {
-  UNCOVERING,
-  FLAG_PLACEMENT,
-  QUESTION_MARK_PLACEMENT,
-  CLEARANCE
+import com.deviget.minesweeper.domain.model.game.cell.CellCoordinate;
+
+public class CellOperation {
+
+  public final CellCoordinate cellCoordinate;
+  public final CellOperationType type;
+  public final long timeElapsed;
+
+  public static CellOperation from(final int rowIndex,
+                            final int cellIndex,
+                            final String type,
+                            final long timeElapsed) {
+    return new CellOperation(CellCoordinate.with(rowIndex, cellIndex), CellOperationType.valueOf(type), timeElapsed);
+  }
+
+  private CellOperation(final CellCoordinate cellCoordinate,
+                        final CellOperationType type,
+                        final long timeElapsed) {
+    this.cellCoordinate = cellCoordinate;
+    this.timeElapsed = timeElapsed;
+    this.type = type;
+  }
 }