+++ src/main/java/com/deviget/minesweeper/infra/service/CellOperationType.java
@@ -1,0 +1,8 @@
+package com.deviget.minesweeper.infra.service;
+
+public enum CellOperationType {
+  UNCOVERING,
+  FLAG_PLACEMENT,
+  QUESTION_MARK_PLACEMENT,
+  CLEARANCE
+}