+++ src/main/java/com/deviget/minesweeper/infra/service/ClearanceOperationStrategy.java
@@ -2,7 +2,6 @@ package com.deviget.minesweeper.infra.service;
 
 import com.deviget.minesweeper.domain.model.game.Game;
 import com.deviget.minesweeper.domain.model.game.GameId;
-import com.deviget.minesweeper.domain.model.game.cell.CellCoordinate;
 import com.deviget.minesweeper.domain.model.user.UserId;
 import com.deviget.minesweeper.infra.repository.GameRepository;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -17,9 +16,9 @@ public class ClearanceOperationStrategy implements CellOperationStrategy {
   @Override
   public Game perform(final GameId gameId,
                       final UserId userId,
-                      final CellCoordinate cellCoordinate) {
+                      final CellOperation operation) {
     final Game existingGame = gameRepository.findByIdAndUserId(gameId, userId);
-    existingGame.clearCell(cellCoordinate);
+    existingGame.clearCell(operation.cellCoordinate, operation.timeElapsed);
     return gameRepository.save(existingGame);
   }