+++ src/main/java/com/deviget/minesweeper/infra/service/GameService.java
@@ -4,7 +4,6 @@ import com.deviget.minesweeper.domain.model.game.Game;
 import com.deviget.minesweeper.domain.model.game.GameId;
 import com.deviget.minesweeper.domain.model.game.GameStatus;
 import com.deviget.minesweeper.domain.model.game.Preferences;
-import com.deviget.minesweeper.domain.model.game.cell.CellCoordinate;
 import com.deviget.minesweeper.domain.model.user.UserId;
 import com.deviget.minesweeper.infra.repository.GameRepository;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -19,26 +18,26 @@ public class GameService {
   private CellOperations cellOperations;
 
   @Autowired
+  private GameStatusTransitions gameStatusTransitions;
+
+  @Autowired
   private GameRepository gameRepository;
 
   public Game configure(final Preferences preferences, final UserId userId) {
     return gameRepository.save(Game.configure(preferences, userId));
   }
 
-  public Game resume(final GameId gameId, final UserId userId) {
-    final Game existingGame = gameRepository.findByIdAndUserId(gameId, userId);
-    existingGame.resume();
-    return gameRepository.save(existingGame);
+  public Game changeStatus(final GameId id, final UserId userId, final GameStatusTransition statusTransition) {
+    return gameStatusTransitions.of(statusTransition.reason).changeStatus(id, userId, statusTransition);
   }
 
   public Game performCellOperation(final GameId gameId,
                                    final UserId userId,
-                                   final CellOperation operation,
-                                   final CellCoordinate cellCoordinate) {
-    return cellOperations.of(operation).perform(gameId, userId, cellCoordinate);
+                                   final CellOperation operation) {
+    return cellOperations.of(operation.type).perform(gameId, userId, operation);
   }
 
-  public List<Game> findUnfinished(final UserId userId) {
-    return gameRepository.findByUserIdAndStatusIn(userId, GameStatus.unfinished());
+  public List<Game> findPaused(final UserId userId) {
+    return gameRepository.findByUserIdAndStatusIn(userId, List.of(GameStatus.PAUSED));
   }
 }